Commit14
Delete Detention, add username and password to add staff

+++ Delete/deleteStaffMember.php
@@ -12,6 +12,13 @@
             die('Could not connect: ' . $con->connect_error);
         }
 
+        $sql = "DELETE FROM user WHERE Staff_ID =". $_GET['id'];
+
+        if(!$con->query($sql))
+        {
+            die('Error: '. $con->error);
+        }
+
         $sql = "DELETE FROM staff WHERE ID =". $_GET['id'];
 
         if(!$con->query($sql))
@@ -19,7 +26,7 @@
             die('Error: '. $con->error);
         }
         echo "<p>Staff Member Deleted</p>";
-        echo "<a href='../View/viewStaff.php'>Back to Staff</a>";
+        echo "<a href='../View/Form/viewStaffForm.phtml'>Back to Staff</a>";
 
         $con->close();
